fix: Correct type mismatch for Gmail History ID (BigInteger vs Long)
- Update FetchMessagesResult to use BigInteger for historyId
- Convert BigInteger to Long when storing in database
- Fixes compilation error in GmailApiClient

This resolves the pipeline build error:
"Argument type mismatch: actual type is 'java.math.BigInteger?',
but 'kotlin.Long?' was expected"

+++ data/src/main/java/com/mailflow/data/repository/EmailRepositoryImpl.kt
@@ -51,7 +51,7 @@ class EmailRepositoryImpl @Inject constructor(
                 gmailSyncStateDao.updateSyncState(
                     GmailSyncStateEntity(
                         id = 1,
-                        historyId = historyId,
+                        historyId = historyId.toLong(),
                         lastSyncTimestamp = System.currentTimeMillis()
                     )
                 )
